+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Next version
 
+# 1.5.3
+
+- deploy WBTC, WMATIC, and USDT test tokens on Mumbai testnet
+- Fix: Error in ActivateMangroveOrder script fixed
+- deploy scripts on Mumbai uses chainling oracle for density parameters
+
 # 1.5.2
 
 - deploy new Mangrove and periphery contracts to Polygon